wgpu: Refactor surface command target to its own file
