+++ render/wgpu/src/surface.rs
@@ -1,151 +1,14 @@
-use crate::buffer_pool::{PoolEntry, TexturePool};
-use crate::commands::{CommandRenderer, CommandTarget};
+pub mod target;
+
+use crate::buffer_pool::TexturePool;
+use crate::commands::CommandRenderer;
 use crate::mesh::Mesh;
 use crate::uniform_buffer::BufferStorage;
 use crate::utils::remove_srgb;
 use crate::{Descriptors, Pipelines, TextureTransforms, Transforms, UniformBuffer};
 use ruffle_render::commands::CommandList;
 use std::sync::Arc;
-
-#[derive(Debug)]
-pub struct ResolveBuffer {
-    texture: PoolEntry<wgpu::Texture>,
-    view: wgpu::TextureView,
-}
-
-impl ResolveBuffer {
-    pub fn new(
-        descriptors: &Descriptors,
-        size: wgpu::Extent3d,
-        format: wgpu::TextureFormat,
-        usage: wgpu::TextureUsages,
-        pool: &mut TexturePool,
-    ) -> Self {
-        let texture = pool.get_texture(descriptors, size, usage, format, 1);
-
-        let view = texture.create_view(&Default::default());
-
-        Self { texture, view }
-    }
-
-    pub fn view(&self) -> &wgpu::TextureView {
-        &self.view
-    }
-
-    pub fn texture(&self) -> &wgpu::Texture {
-        &self.texture
-    }
-
-    pub fn take_texture(self) -> PoolEntry<wgpu::Texture> {
-        self.texture
-    }
-}
-
-#[derive(Debug)]
-pub struct FrameBuffer {
-    texture: PoolEntry<wgpu::Texture>,
-    view: wgpu::TextureView,
-    size: wgpu::Extent3d,
-}
-
-impl FrameBuffer {
-    pub fn new(
-        descriptors: &Descriptors,
-        sample_count: u32,
-        size: wgpu::Extent3d,
-        format: wgpu::TextureFormat,
-        usage: wgpu::TextureUsages,
-        pool: &mut TexturePool,
-    ) -> Self {
-        let texture = pool.get_texture(descriptors, size, usage, format, sample_count);
-
-        let view = texture.create_view(&Default::default());
-
-        Self {
-            texture,
-            view,
-            size,
-        }
-    }
-
-    pub fn view(&self) -> &wgpu::TextureView {
-        &self.view
-    }
-
-    pub fn texture(&self) -> &wgpu::Texture {
-        &self.texture
-    }
-
-    pub fn take_texture(self) -> PoolEntry<wgpu::Texture> {
-        self.texture
-    }
-
-    pub fn size(&self) -> wgpu::Extent3d {
-        self.size
-    }
-}
-
-#[derive(Debug)]
-pub struct BlendBuffer {
-    texture: PoolEntry<wgpu::Texture>,
-    view: wgpu::TextureView,
-}
-
-impl BlendBuffer {
-    pub fn new(
-        descriptors: &Descriptors,
-        size: wgpu::Extent3d,
-        format: wgpu::TextureFormat,
-        usage: wgpu::TextureUsages,
-        pool: &mut TexturePool,
-    ) -> Self {
-        let texture = pool.get_texture(descriptors, size, usage, format, 1);
-        let view = texture.create_view(&Default::default());
-
-        Self { texture, view }
-    }
-
-    pub fn view(&self) -> &wgpu::TextureView {
-        &self.view
-    }
-
-    pub fn texture(&self) -> &wgpu::Texture {
-        &self.texture
-    }
-}
-
-#[derive(Debug)]
-pub struct DepthBuffer {
-    _texture: PoolEntry<wgpu::Texture>,
-    view: wgpu::TextureView,
-}
-
-impl DepthBuffer {
-    pub fn new(
-        descriptors: &Descriptors,
-        msaa_sample_count: u32,
-        size: wgpu::Extent3d,
-        pool: &mut TexturePool,
-    ) -> Self {
-        let texture = pool.get_texture(
-            descriptors,
-            size,
-            wgpu::TextureUsages::RENDER_ATTACHMENT,
-            wgpu::TextureFormat::Depth24PlusStencil8,
-            msaa_sample_count,
-        );
-
-        let view = texture.create_view(&Default::default());
-        Self {
-            _texture: texture,
-            view,
-        }
-    }
-
-    pub fn view(&self) -> &wgpu::TextureView {
-        &self.view
-    }
-}
+use target::CommandTarget;
 
 #[derive(Debug)]
 pub struct Surface {
